+++ Observer/Sales/Order/Invoice/AfterSave.php
@@ -23,6 +23,7 @@ class AfterSave implements ObserverInterface
 //    private $paymentRequestFactory;
 //    private $requestFactory;
     private $readerWriter;
+//    private $invRepo;
     
     public function __construct(
 //        \Nuvei\Checkout\Model\Config $config,
@@ -31,13 +32,15 @@ class AfterSave implements ObserverInterface
         \Magento\Framework\ObjectManagerInterface $objectManager,
 //        \Magento\Framework\Controller\Result\JsonFactory $jsonResultFactory,
         \Nuvei\Checkout\Model\ReaderWriter $readerWriter
+//        \Magento\Sales\Model\Order\InvoiceRepository $invRepo
     ) {
 //        $this->config                   = $config;
 //        $this->paymentRequestFactory    = $paymentRequestFactory;
 //        $this->requestFactory           = $requestFactory;
-        $this->objectManager            = $objectManager;
+        $this->objectManager    = $objectManager;
 //        $this->jsonResultFactory        = $jsonResultFactory;
-        $this->readerWriter             = $readerWriter;
+        $this->readerWriter     = $readerWriter;
+//        $this->invRepo          = $invRepo;
     }
     
     /**
@@ -47,6 +50,8 @@ class AfterSave implements ObserverInterface
      */
     public function execute(Observer $observer)
     {
+        $this->readerWriter->createLog('Invoice AfterSave Observer');
+        
         try {
             /** @var Invoice $invoice */
             $invoice = $observer->getInvoice();
@@ -92,13 +97,13 @@ class AfterSave implements ObserverInterface
             }
 
             // Settle request
-            $authCode                = '';
-            $ord_trans_addit_info    = $payment->getAdditionalInformation(Payment::ORDER_TRANSACTIONS_DATA);
+            $authCode               = '';
+            $inv_id                 = $invoice->getId();
+            $ord_trans_addit_info   = $payment->getAdditionalInformation(Payment::ORDER_TRANSACTIONS_DATA);
             
             // probably a Sale
             if (!is_array($ord_trans_addit_info)
                 || empty($ord_trans_addit_info)
-                || count($ord_trans_addit_info) < 1
             ) {
                 return $this;
             }
@@ -106,12 +111,12 @@ class AfterSave implements ObserverInterface
             foreach ($ord_trans_addit_info as $trans) {
                 if (strtolower($trans[Payment::TRANSACTION_STATUS]) == 'approved') {
                     if (strtolower($trans[Payment::TRANSACTION_TYPE]) == 'sale') {
-                        $this->readerWriter->createLog('After Save Invoice observer - Sale');
+//                        $this->readerWriter->createLog('After Save Invoice observer - Sale.');
                         return $this;
                     }
 
                     if (strtolower($trans[Payment::TRANSACTION_TYPE]) == 'auth') {
-                        $this->readerWriter->createLog('After Save Invoice observer - Auth');
+//                        $this->readerWriter->createLog('After Save Invoice observer - Auth.');
                         $authCode = $trans[Payment::TRANSACTION_AUTH_CODE];
                         break;
                     }
@@ -130,16 +135,28 @@ class AfterSave implements ObserverInterface
             
             $request = $this->objectManager->create(\Nuvei\Checkout\Model\Request\SettleTransaction::class);
 
-            $request
+            $resp = $request
                 ->setPayment($payment)
                 ->setInvoiceId($invoice->getId())
                 ->setInvoiceAmount($invoice->getBaseGrandTotal())
                 ->process();
+            
+            
+//            if(empty($resp['transactionStatus']) || 'APPROVED' != $resp['transactionStatus']) {
+////                $invoice->setState(Invoice::STATE_CANCELED);
+////                $this->invRepo->save($invoice);
+//                
+//                $this->readerWriter->createLog('Invoice AfterSave try to delete the invoice');
+//                $invoice_data = $this->invRepo->get($inv_id);
+//                
+//                $this->invRepo->delete($invoice_data);
+//            }
             // Settle request END
         } catch (Exception $e) {
             $this->readerWriter->createLog($e->getMessage(), 'Invoice AfterSave Exception');
         }
         
+//        $this->readerWriter->createLog('End of AfterSave Invoice observer.');
         return $this;
     }
 }