+++ Observer/Sales/Order/Invoice/Register.php
@@ -65,13 +65,6 @@ class Register implements ObserverInterface
         $invoice    = $observer->getInvoice();
         $inv_state  = Invoice::STATE_OPEN; // in case of auth we will change it when DMN come
         
-        $ord_trans_addit_info = $payment->getAdditionalInformation(Payment::ORDER_TRANSACTIONS_DATA);
-        
-        // in case of Sale
-        if (!is_array($ord_trans_addit_info) || count($ord_trans_addit_info) < 1) {
-            $inv_state  = Invoice::STATE_PAID;
-        }
-        
         $invoice->setState($inv_state);
         
         return $this;